pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "acronym-cli"
version = "0.1.0"
description = "Command-line tool to manage acronyms with default lookup"
authors = [
    { name="Paddy", email="you@example.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"

dependencies = [
    "click>=8.0",
    "PyYAML",
    "rich"
]

[project.scripts]
acronym = "acronym.cli:main"

